Updated DeepestLSTM training procedure
