+++ Formula1-FollowLine/tensorflow/DeepestLSTMTinyPilotNet/utils/deepest_lstm_tinypilotnet.py
@@ -1,8 +1,9 @@
 from tensorflow.keras.models import Sequential
-from tensorflow.keras.layers import Flatten,Dense,Conv2D,BatchNormalization,Dropout,ConvLSTM2D,Reshape,Activation,MaxPooling2D, LSTM, Input
+from tensorflow.keras.layers import Flatten, Dense, Conv2D, BatchNormalization, Dropout, ConvLSTM2D, Reshape, \
+    Activation, MaxPooling2D, LSTM, Input
 from tensorflow.keras.optimizers import Adam
 
-# PAPER Self-driving a Car in Simulation Through a CNN
+
 # DEEPEST LSTM tinypilotnet
 def deepest_lstm_tinypilotnet_model(img_shape):
     model = Sequential()
